% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_ids_pred_presencia.R
\name{llamada_prediccion_presencia_temporal}
\alias{llamada_prediccion_presencia_temporal}
\title{Llama N veces a la funcion de prediccion presencia donde N es el numero de dispositivos relacionados con el activo recbido}
\usage{
llamada_prediccion_presencia_temporal(
  fecha_inicial,
  fecha_final,
  tipo_estancia
)
}
\arguments{
\item{fecha_inicial, }{fecha_final, tipo_estancia}
}
\value{
json
}
\description{
Llama N veces a la funcion de prediccion presencia donde N es el numero de dispositivos relacionados con el activo recbido
}
\examples{
 llamada_prediccion_presencia_temporal("2021-07-10 00:00:00", "2021-08-10 23:00:00", "aulas-pb")

}
