% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_presencia_temp_knn_func.R
\name{prediccion_presencia_temporal}
\alias{prediccion_presencia_temporal}
\title{Predice la existencia de presencia en una estancia objetivo en los 2 registros anteriores relativos a la marca temporal}
\usage{
prediccion_presencia_temporal(
  fecha_inicial = as.character(Sys.time() - 30 * 24 * 60 * 60),
  fecha_final = as.character(Sys.time()),
  id_estancia = "f4d0d070-24dd-11eb-b605-01af9c6dd825",
  tipo_estancia = "despachos-pb"
)
}
\arguments{
\item{fecha_inicial, }{fecha_final, id_estancia, tipo_estancia}
}
\value{
json
}
\description{
Predice la existencia de presencia en una estancia objetivo en los 2 registros anteriores relativos a la marca temporal.
Recibe los datos de condiciones ambientales de la estancia objetivo del último mes.
Procesa, limpia y prepara los datos para la predicción.
Utiliza el algoritmo knn para predecir los últimos dos registro en la serie de tiempo.
Estos registros marcarán la existencia de presencia en la sala en el t+30' donde t es la marca temporal del último registro.
Para que exista presencia, el resultado de la predicción de los dos últimos dos registros debe ser de 1 (presencia).
tipo_estancia: aulas-pb, despachos-pb, empresas-pb, seminarios-pb, despachos-p1
}
\examples{
 prediccion_presencia_temporal("2021-07-10 00:00:00", "2021-08-10 23:00:00", "03179970-24de-11eb-b605-01af9c6dd825", "aulas-pb")

}
